move editor config to new file
