+++ src/index.ts
@@ -5,12 +5,11 @@ import Piece from "./piece";
 import { saveSelection, restoreSelection } from "./utils/selectionManager";
 import { parseHtmlToPieces } from "./utils/parseHtml";
 import { showHyperlinkViewButton, hideHyperlinkViewButton } from './attributes/hyperLink'
+import { createEditor } from "./config/editorConfig";
 import "./styles/text-igniter.css"
-import { icons } from "./assets/icons";
 
-export type EditorConfig = {
-    features: string[]
-}
+import { EditorConfig } from "./types/editorConfig";
+
 
 export interface CurrentAttributeDTO { bold: boolean; italic: boolean; underline: boolean; undo?: boolean; redo?: boolean, hyperlink?: string | boolean, fontFamily?: string; fontSize?: string; }
 
@@ -27,10 +26,10 @@ class TextIgniter {
 
     constructor(editorId: string, config: EditorConfig) {
 
-        this.createEditor(editorId, config);
+        const{mainEditorId,toolbarId} = createEditor(editorId, config);
 
-        this.editorContainer = document.getElementById('editor') || null;
-        this.toolbarContainer = document.getElementById('toolbar') || null;
+        this.editorContainer = document.getElementById(mainEditorId) || null;
+        this.toolbarContainer = document.getElementById(toolbarId) || null;
 
         if (!this.editorContainer || !this.toolbarContainer) {
             throw new Error("Editor element not found or incorrect element type.");
@@ -222,200 +221,6 @@ class TextIgniter {
 
     }
 
-    createEditor(editorId: string, config: EditorConfig) {
-        const allowedFontFamily = [
-            'Arial',
-            'Times New Roman',
-            'Courier New',
-            'Verdana',
-        ];
-        const allowedFontSizes = [
-            '12px',
-            '14px',
-            '16px',
-            '18px',
-            '20px',
-        ];
-        const container = document.getElementById(editorId);
-        if (!container) {
-            throw new Error("Editor element not found or incorrect element type.");
-        }
-        const toolbar = document.createElement('div');
-        toolbar.className = 'toolbar';
-        toolbar.id = 'toolbar';
-        container.appendChild(toolbar);
-
-        const editor = document.createElement('div');
-        editor.id = 'editor';
-        editor.contentEditable = 'true';
-        container.appendChild(editor);
-
-        // Map features to button labels/icons
-        const featureLabels: any = {
-            'bold': '<strong>B</strong>',
-            'italic': '<em>I</em>',
-            'underline': '<u>U</u>',
-            'hyperlink': '&#128279;',   // Unicode for link symbol
-
-            'alignLeft': '&#8676;',    // Unicode for left arrow
-            'alignCenter': '&#8596;',  // Unicode for left-right arrow
-            'alignRight': '&#8677;',   // Unicode for right arrow
-
-            'unorderedList': '&#8226;',   // Unicode for bullet
-            'orderedList': '1.',      // Simple text representation
-            'fontFamily': 'fontFamily',
-            'fontSize': 'fontSize',
-
-            'subscript': 'X<sub>2</sub>',
-            'superscript': 'X<sup>2</sup>',
-            'justify': '&#8644;',       // Unicode for justify icon
-            'insert_table': '&#8866;',  // Unicode for table icon
-            'insert_layout': '&#10064;',// Unicode for layout icon
-            'heading': 'H',
-            'image': '&#128247;',       // Unicode for camera symbol
-            'colors': '&#127912;',      // Unicode for palette symbol
-        };
-
-        // Features with custom SVG icons
-        const featuresWithPngIcon = [
-            { feature: 'alignLeft', id: 'alignLeft', icon: icons.left_align },
-            { feature: 'alignCenter', id: 'alignCenter', icon: icons.center_align },
-            { feature: 'alignRight', id: 'alignRight', icon: icons.right_align },
-            { feature: 'unorderedList', id: 'unorderedList', icon: icons.bullet_list },
-            { feature: 'orderedList', id: 'orderedList', icon: icons.numbered_list },
-        ];
-
-        const createSelect = (id: string, options: string[]) => {
-            const select = document.createElement('select');
-            select.dataset.action = id;
-            select.id = id;
-            options.forEach(optionValue => {
-                const option = document.createElement('option');
-                option.value = optionValue;
-                option.textContent = optionValue;
-                select.appendChild(option);
-            });
-            return select;
-        };
-
-        config.features.forEach(feature => {
-            if (feature === 'fontFamily') {
-                const fontFamilySelect = createSelect('fontFamily', allowedFontFamily);
-                toolbar.appendChild(fontFamilySelect);
-            } else if (feature === 'fontSize') {
-                const fontSizeSelect = createSelect('fontSize', allowedFontSizes);
-                toolbar.appendChild(fontSizeSelect);
-            } else if (featuresWithPngIcon.map(item => item.feature).indexOf(feature) !== -1) {
-                const featureDataArray = featuresWithPngIcon.filter(item => item.feature === feature);
-                let featureData = null;
-                if (featureDataArray?.length > 0) {
-                    featureData = featureDataArray[0];
-                }
-                const button = document.createElement('button');
-                button.id = feature;
-                button.dataset.action = feature;
-                const svg = featureData?.icon || "";
-                button.innerHTML = svg;
-                toolbar.appendChild(button);
-
-                // Commented for future use
-
-                // const img = document.createElement('img');
-                // img.src = featureData?.icon || "";
-                // img.width = 20;
-                // img.height = 20;
-                // button.appendChild(img);
-                toolbar.appendChild(button);
-            } else {
-                const button = document.createElement('button');
-                button.dataset.action = feature;
-                button.innerHTML = featureLabels[feature] || feature;
-                button.id = feature;
-                // if(['leftAlign','centerAlign','rightAlign','bulletList','numberedList'].indexOf(feature) !== -1){
-                //     button.id = feature;
-                // }
-                // Add the title attribute for hover effect
-                button.title = feature
-                    .split('_')
-                    .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-                    .join(' ');
-
-                toolbar.appendChild(button);
-            }
-
-        });
-
-        // Create the container div
-        const hyperlinkContainer = document.createElement("div");
-        hyperlinkContainer.id = "hyperlink-container";
-        hyperlinkContainer.style.display = "none";
-
-        // Create the input element
-        const hyperlinkInput = document.createElement("input");
-        hyperlinkInput.type = "text";
-        hyperlinkInput.id = "hyperlink-input";
-        hyperlinkInput.placeholder = "Enter a URL...";
-
-        // Create the Apply button
-        const applyButton = document.createElement("button");
-        applyButton.id = "apply-hyperlink";
-        applyButton.textContent = "Link";
-
-        // Create the Cancel button
-        const cancelButton = document.createElement("button");
-        cancelButton.id = "cancel-hyperlink";
-        cancelButton.textContent = "Unlink";
-
-        // Append input and buttons to the container
-        hyperlinkContainer.appendChild(hyperlinkInput);
-        hyperlinkContainer.appendChild(applyButton);
-        hyperlinkContainer.appendChild(cancelButton);
-
-        // Append the container to the toolbar
-
-        toolbar.appendChild(hyperlinkContainer);
-
-
-
-
-        // Create the container div
-        const viewHyperlinkContainer = document.createElement("div");
-        viewHyperlinkContainer.id = "hyperlink-container-view";
-        viewHyperlinkContainer.style.display = "none";
-
-        //  // Create the input element
-        const hyperLinkViewSpan = document.createElement("span");
-        hyperLinkViewSpan.id = "hyperlink-view-span";
-        hyperLinkViewSpan.innerHTML = "Visit URL : ";
-
-        const hyperLinkAnchor = document.createElement("a");
-        hyperLinkAnchor.id = "hyperlink-view-link";
-        hyperLinkAnchor.href = "";
-        hyperLinkAnchor.target = "_blank";
-
-
-        // Create the Apply button
-        // const editHyperlinkButton = document.createElement("button");
-        // editHyperlinkButton.id = "edit-hyperlink";
-        // editHyperlinkButton.textContent = "edit |";
-
-        // Create the Cancel button
-        // const removeHyperlinkButton = document.createElement("button");
-        // removeHyperlinkButton.id = "delete-hyperlink";
-        // removeHyperlinkButton.textContent = "remove";
-
-
-        //  // Append input and buttons to the container
-        viewHyperlinkContainer.appendChild(hyperLinkViewSpan);
-        viewHyperlinkContainer.appendChild(hyperLinkAnchor);
-        //  viewHyperlinkContainer.appendChild(editHyperlinkButton);
-        //  viewHyperlinkContainer.appendChild(removeHyperlinkButton);
-
-        //  // Append the container to the toolbar
-
-        toolbar.appendChild(viewHyperlinkContainer);
-    }
-
     getSelectionRange(): [number, number] {
         const sel = saveSelection(this.editorView.container);
         if (!sel) return [0, 0];
@@ -515,16 +320,8 @@ class TextIgniter {
                     this.document.redo();
                     break;
                 case 'hyperlink':
-                    if (start < end) {
-                        // Get the existing hyperlink, if any
                         const existingLink = this.document.getCommonHyperlinkInRange(start, end);
-
-                        // Show the hyperlink input box
                         this.showHyperlinkInput(existingLink);
-                    } else {
-                        // No selection
-                        alert("Please select the text you want to hyperlink.");
-                    }
                     break;
             }
         } else {
@@ -575,14 +372,13 @@ class TextIgniter {
                 const url = hyperlinkInput.value.trim();
                 if (url) {
                     this.applyHyperlink(url);
-                } else {
-                    // alert('Please enter a valid URL.');
                 }
                 hyperlinkContainer.style.display = 'none';
             };
 
             // Handle the 'Unlink' button
             cancelButton.onclick = () => {
+        console.log('zzz',{remove:this.document.dataIds})
                 this.removeHyperlink();
                 hyperlinkContainer.style.display = 'none';
             };
@@ -635,7 +431,30 @@ class TextIgniter {
 
         const [start, end] = this.getSelectionRange();
         if (start < end) {
-            this.document.applyHyperlinkRange(start, end, url);
+            console.log('zzz',this.document.blocks)
+            console.log('zzz',this.document.selectedBlockId)
+            console.log('zzz',this.document.dataIds)
+            
+            if (this.document.dataIds.length > 1) {
+                this.document.blocks.forEach((block: any) => {
+                    if (this.document.dataIds.includes(block.dataId)) {
+                        this.document.selectedBlockId = block.dataId;
+                        let countE = 0;
+                        block.pieces.forEach((obj: any) => {
+                            countE += obj.text.length;
+                        })
+                        let countS = start - countE;
+                        this.document.applyHyperlinkRange(countS, countE,url);
+                    }
+                })
+            } else {
+                this.document.applyHyperlinkRange(start, end,url);
+            }
+          
+
+
+            // this.document.applyHyperlinkRange(start, end, url);
+
             this.editorView.render();
             // Restore selection and focus
             restoreSelection(this.editorView.container, this.savedSelection);
@@ -652,12 +471,42 @@ class TextIgniter {
         restoreSelection(this.editorView.container, this.savedSelection);
 
         const [start, end] = this.getSelectionRange();
+        console.log('zzz',{remove:this.document.dataIds})
         if (start < end) {
-            this.document.removeHyperlinkRange(start, end);
+            // this.document.removeHyperlinkRange(start, end);
+
+
+
+
+
+
+            if (this.document.dataIds.length > 1) {
+                this.document.blocks.forEach((block: any) => {
+                    if (this.document.dataIds.includes(block.dataId)) {
+                        this.document.selectedBlockId = block.dataId;
+                        let countE = 0;
+                        block.pieces.forEach((obj: any) => {
+                            countE += obj.text.length;
+                        })
+                        let countS = start - countE;
+                        this.document.removeHyperlinkRange(countS, countE);
+                    }
+                })
+            } else {
+                this.document.removeHyperlinkRange(start, end);
+            }
+          
+
+
+
             this.editorView.render();
             // Restore selection and focus
             restoreSelection(this.editorView.container, this.savedSelection);
             this.editorView.container.focus();
+
+
+
+
         }
         this.savedSelection = null;
     }
@@ -681,7 +530,7 @@ class TextIgniter {
         const [start, end] = this.getSelectionRange();
         let ending = end;
         if (e.key === 'Enter') {
-            console.log('blocks', this.document.blocks)
+            console.log('blocks--->>', this.document.blocks)
             e.preventDefault();
             const uniqueId = `data-id-${Date.now()}`;
             if (this.document.blocks[this.document.blocks.length - 1]?.listType === 'ol' || this.document.blocks[this.document.blocks.length - 1]?.listType === 'ul' || this.document.blocks[this.document.blocks.length - 1]?.listType === 'li') {
@@ -713,17 +562,21 @@ class TextIgniter {
                     listStart: ListType === 'ol' || ListType === 'li' ? _start : '',
                 })
             } else {
+                console.log('jagdiii 0')
+
                 console.log('vk11', this.getCurrentCursorBlock())
                 if (this.getCurrentCursorBlock() !== null) {
                     const extractedContent = " " + this.extractTextFromDataId(this.getCurrentCursorBlock()!.toString())
                     console.log("vk11, ", this.getCurrentCursorBlock()!.toString(), " - ", start, end, extractedContent)
                     let updatedBlock = this.document.blocks;
-
+                    console.log({extractedContent})
                     if (extractedContent.length > 0) {
                         updatedBlock = this.addBlockAfter(this.document.blocks, this.getCurrentCursorBlock()!.toString(), {
                             "dataId": uniqueId, "class": "paragraph-block", "pieces": [new Piece(extractedContent)],
+                            // "dataId": uniqueId, "class": "paragraph-block", "pieces": [new Piece(" ")],
                             // listType: null, // null | 'ol' | 'ul'
                         });
+                        console.log('schenerio1');
                         ending = start + extractedContent.length - 1;
                     } else {
                         updatedBlock = this.addBlockAfter(this.document.blocks, this.getCurrentCursorBlock()!.toString(), {
@@ -735,6 +588,7 @@ class TextIgniter {
                     this.document.blocks = updatedBlock
                     console.log("vk11", this.document.blocks, " updatedBlock", updatedBlock)
                 } else {
+                    console.log('jagdiii 1')
                     this.document.blocks.push({
                         "dataId": uniqueId, "class": "paragraph-block", "pieces": [new Piece(" ")],
                         // listType: null, // null | 'ol' | 'ul'
@@ -839,9 +693,14 @@ class TextIgniter {
         // Get the full text content of the element
         const fullText = element.textContent || '';
 
+        console.log("oooo",{fullText});
         // Calculate the offset position of the cursor within the text node
         const cursorOffset = range.startOffset;
+        const textLengthTillCurrentCursor = this.document.blocks.filter((block:any) => block.dataId === dataId)[0].pieces.reduce((acc:number,currVal:Piece) =>acc+currVal.text.length,0);
+        console.log("oooo",{textLengthTillCurrentCursor})
 
+        console.log("oooo",{cursorOffset})
+        console.log("oooo",{end:range.endOffset});
         // Extract text from the cursor position to the end
         const remainingText = fullText.slice(cursorOffset);
 
@@ -894,6 +753,7 @@ class TextIgniter {
     }
     syncCurrentAttributesWithCursor(): void {
         const [start, end] = this.getSelectionRange();
+        console.log("oooo",{start});
         if (start === end) {
             const piece = this.document.findPieceAtOffset(start, this.document.selectedBlockId);
             if (piece) {
