+++ src/types/editorConfig.ts
@@ -1,0 +1,8 @@
+export type EditorConfig = {
+    features: string[]
+}
+
+export type EditorConfigReturnType = {
+     mainEditorId: string;
+      toolbarId: string 
+}
